refactor: Refactor large getStaticProps statements into functions
