+++ pages/index.tsx
@@ -5,12 +5,12 @@ import FeaturedRepositories from "components/FeaturedRepositories";
 import BlogPosts from "components/BlogPosts";
 import Snippets from "components/Snippets";
 import { getClient } from "lib/sanity-server";
-import { limitedSnippetsQuery, threePostsQuery } from "lib/queries";
+import { limitedSnippetsQuery } from "lib/queries";
 import { Snippet, Post } from "lib/types";
-import prisma from "lib/prisma";
 import PageWidth from "components/layout/PageWidth";
 import Section from "components/home/Section";
 import Landing from "components/home/Landing";
+import getPopularPosts from "utils/getPopularPosts";
 
 interface Props {
   snippets: Snippet[];
@@ -47,28 +47,12 @@ export const getStaticProps: GetStaticProps = async ({ preview = false }) => {
     }
   );
 
-  const topThreeSlugs = await (
-    await prisma.blogViews.findMany({ orderBy: [{ count: "desc" }], take: 3 })
-  ).map((post) => post.slug);
-
-  const posts: Post[] = await getClient(preview).fetch(threePostsQuery, {
-    slug_1: topThreeSlugs[0] || "",
-    slug_2: topThreeSlugs[1] || "",
-    slug_3: topThreeSlugs[2] || "",
-  });
-
-  const orderedPosts = posts
-    .map((_post, index) => {
-      const slug = topThreeSlugs[index];
-      const post = posts.find((post) => post.slug === slug);
-      return post;
-    })
-    .filter((post) => !!post);
+  const posts: Post[] = await getPopularPosts();
 
   return {
     props: {
-      snippets: snippets || [],
-      posts: orderedPosts || [],
+      snippets,
+      posts,
     },
   };
 };
