+++ pages/snippets/[slug].tsx
@@ -1,5 +1,6 @@
 import { MDXRemote } from "next-mdx-remote";
 import { GetStaticPaths, GetStaticProps } from "next";
+
 import SnippetLayout from "layouts/snippet";
 import { snippetsQuery, snippetSlugsQuery } from "lib/queries";
 import { sanityClient, getClient } from "lib/sanity-server";
@@ -22,7 +23,10 @@ export const getStaticPaths: GetStaticPaths = async () => {
   };
 };
 
-export const getStaticProps: GetStaticProps = async ({ params, preview = false }) => {
+export const getStaticProps: GetStaticProps = async ({
+  params,
+  preview = false,
+}) => {
   const { snippet } = await getClient(preview).fetch(snippetsQuery, {
     slug: params?.slug,
   });
